Added some POC code to allow a custom go build
By default this is fairly tight in terms of how the go build command works. I
need it to be a little more flexible, so this can now be done using the -c
(--customBuild) flag when running Gin.

For example:

gin -c "go build -o hello -a -tags netgo -ldflags '-w' ."

This probably doesn't have a ton of use cases for this particular package but
I need it myself :)